Don't extend `String` since it's really `StringType` at runtime (#1)
In order to make microstates convenient, `Number`, `String`, `Boolean`
are actually standins for the microstate types `NumberType`,
`StringType`, and `BooleanType` when specifying them as child
properties:

```javascript
class MyThing {
  num = Number
  str = String
}
```

this gets expanded to:

```javascript
class MyThing {
  num = NumberType
  str = StringType
}
```

We can't do the same thing (or at least not very easily) with
inheritance, so what was happening was that when the generated field
`class Field extends String` constructor was being invoked, it was
invoking the super String constructor which returned an instance of
`String` "undefined", or "[object Object]" depending on the value that
got passed in. This in turn was messing up the child lookup because
microstates thought there was a child for each character in the
string.

This removes the inheritance (maybe containment is a better strategy)
although it could be added back by exmplicitly `import { StringType }
from 'microstates'` and descending from _that_ class.